Webhook refactor
