+++ apps/products-feed/src/pages/api/webhooks/product_deleted.ts
@@ -8,6 +8,7 @@ import { updateCacheForConfigurations } from "../../../modules/metadata-cache/up
 import { saleorApp } from "../../../saleor-app";
 import { createLogger } from "@saleor/apps-shared";
 import { createClient } from "../../../lib/create-graphq-client";
+import { updateCacheOnWebhook } from "../../../modules/metadata-cache/update-cache-on-webhook";
 
 export const config = {
   api: {
@@ -32,30 +33,11 @@ export const handler: NextWebhookApiHandler<ProductWebhookPayloadFragment> = asy
   res,
   context
 ) => {
-  const { event, authData, payload } = context;
-
-  const client = createClient(authData.saleorApiUrl, async () =>
-    Promise.resolve({ token: authData.token })
-  );
-
-  const channelsSlugs = [
-    payload.channel,
-    ...(payload.channelListings?.map((cl) => cl.channel.slug) ?? []),
-  ].filter((c) => c) as string[];
-
-  if (channelsSlugs.length === 0) {
-    res.status(200).end();
-    return;
-  }
-
-  await updateCacheForConfigurations({
-    channelsSlugs,
-    client,
-    saleorApiUrl: authData.saleorApiUrl,
+  await updateCacheOnWebhook({
+    authData: context.authData,
+    channels: context.payload,
+    res,
   });
-
-  res.status(200).end();
-  return;
 };
 
 export default webhookProductDeleted.createHandler(handler);
